backend/requirements.txt
# ============================================================
# CORE FASTAPI + SERVER
# ============================================================
fastapi==0.115.2
uvicorn[standard]==0.30.5
aiosqlite==0.20.0
sqlmodel>=0.0.16

# Required for Jinja2Templates
Jinja2>=3.1.4
MarkupSafe>=2.1.3

# Required for HTML form handling in FastAPI
python-multipart>=0.0.9

# ============================================================
# DATABASE + ORM
# ============================================================
sqlalchemy==2.0.34 
alembic==1.13.2 
asyncpg==0.29.0 
psycopg2-binary==2.9.9 
python-jose==3.3.0

# ============================================================
# Pydantic & Validation
# ============================================================
pydantic==2.8.2 
pydantic-settings==2.3.4 

# ============================================================
# SCHEDULING / BACKGROUND TASKS (APScheduler)
# ============================================================
APScheduler==3.10.4 
# ============================================================
# LOGGING + UTILITIES
# ============================================================
python-dotenv==1.0.1 

# ============================================================
# DEV MODE (SAFE, lightweight)
# These are safe for now and needed for interactive API docs & dev workflow.
# ============================================================
httpx==0.27.0  
pytest==8.3.3   
pytest-asyncio==0.23.6  
# ============================================================
# SECURITY (OPTIONAL but recommended)
# ============================================================
bcrypt==4.2.0  
passlib[bcrypt]==1.7.4  

# ============================================================
# OPTIONAL UTILITIES (KEEP – future use)
# ============================================================
python-multipart==0.0.9 
email-validator==2.1.1

